Obstacles simulation, changed greedy step to use floats for distance comparisons and a-star to use manhatten again

+++ src/greedy_step.h
@@ -13,7 +13,7 @@
  *                  and not in between them, for ease of implementation.
  */
 typedef struct neighbour {
-    int distance;
+    float distance;
     direction_e direction;
     int x;
     int y;
